Import with IntelliJ
impl_test has the following dependencies in (Project Settings -> Modules -> impl_test -> Dependencies):
  - impl_main (correct)
  - <repository>/impl/build/classes/java/main (incorrect, this is a Gradle controlled directory)

IntelliJ uses separate build output directories, so this causes duplicate classes to be found on the
classpath.  Only impl_main, the IntelliJ module, should be present as a dependency.

If the Groovy plugin is removed from the project configuration, the directory disappears.